update code to work with mainframe update

+++ include/bt/api/class/maptech.h
@@ -6,7 +6,7 @@
 namespace bt {
 	class ApiMapTech {
 	public:
-		size_t level;
+		int level;
 		float value;
 	};
 	void from_json(const nlohmann::json& j, ApiMapTech& obj);
